Task #5: Introduced huge refactoring. Main function is now only responsible for the state machine. Created a new class for the CNN, distinguished classes and functions into separated folders. Improved namings and added more comments.
