+++ .gitignore
@@ -1,2 +1,3 @@
 logs/
-videos/
+venv/
+.idea/
